CSP-469 - Operator tests moved to correctly structured test area
