+++ src/SFA.DAS.AANHub.Application.UnitTests/Partners/Commands/CreatePartnersMemberCommandValidatorTests.cs
@@ -8,9 +8,9 @@ namespace SFA.DAS.AANHub.Application.UnitTests.Partners.Commands
 {
     public class CreatePartnerMemberCommandValidatorTests
     {
-        private readonly Mock<IRegionsReadRepository> _regionsReadRepository;
         private readonly Mock<IMembersReadRepository> _membersReadRepository;
         private readonly Mock<IPartnersReadRepository> _partnersReadRepository;
+        private readonly Mock<IRegionsReadRepository> _regionsReadRepository;
 
         public CreatePartnerMemberCommandValidatorTests()
         {
@@ -25,8 +25,11 @@ namespace SFA.DAS.AANHub.Application.UnitTests.Partners.Commands
         [TestCase(" ", false)]
         public async Task Validates_UserName_NotNullOrEmpty(string? userName, bool isValid)
         {
+            var command = new CreatePartnerMemberCommand
+            {
+                UserName = userName!
+            };
 
-            var command = new CreatePartnerMemberCommand() { UserName = userName! };
             var sut = new CreatePartnerMemberCommandValidator(_regionsReadRepository.Object, _membersReadRepository.Object, _partnersReadRepository.Object);
 
             var result = await sut.TestValidateAsync(command);
@@ -36,11 +39,16 @@ namespace SFA.DAS.AANHub.Application.UnitTests.Partners.Commands
             else
                 result.ShouldHaveValidationErrorFor(c => c.UserName);
         }
+
         [TestCase(5, true)]
         [TestCase(251, false)]
         public async Task Validates_UserName_Length(int length, bool isValid)
         {
-            var command = new CreatePartnerMemberCommand() { UserName = new string('a', length) };
+            var command = new CreatePartnerMemberCommand
+            {
+                UserName = new string('a', length)
+            };
+
             var sut = new CreatePartnerMemberCommandValidator(_regionsReadRepository.Object, _membersReadRepository.Object, _partnersReadRepository.Object);
             var result = await sut.TestValidateAsync(command);
 
@@ -49,14 +57,18 @@ namespace SFA.DAS.AANHub.Application.UnitTests.Partners.Commands
             else
                 result.ShouldHaveValidationErrorFor(c => c.UserName);
         }
+
         [TestCase("Organisation name", true)]
         [TestCase(null, false)]
         [TestCase("", false)]
         [TestCase(" ", false)]
         public async Task Validates_Organisation_NotNull(string? organisation, bool isValid)
         {
+            var command = new CreatePartnerMemberCommand
+            {
+                Organisation = organisation!
+            };
 
-            var command = new CreatePartnerMemberCommand() { Organisation = organisation };
             var sut = new CreatePartnerMemberCommandValidator(_regionsReadRepository.Object, _membersReadRepository.Object, _partnersReadRepository.Object);
 
             var result = await sut.TestValidateAsync(command);
@@ -66,12 +78,16 @@ namespace SFA.DAS.AANHub.Application.UnitTests.Partners.Commands
             else
                 result.ShouldHaveValidationErrorFor(c => c.Organisation);
         }
+
         [TestCase(123, true)]
         [TestCase(251, false)]
         public async Task Validates_Organisation_Length(int stringLength, bool isValid)
         {
+            var command = new CreatePartnerMemberCommand
+            {
+                Organisation = new string('a', stringLength)
+            };
 
-            var command = new CreatePartnerMemberCommand() { Organisation = new string('a', stringLength) };
             var sut = new CreatePartnerMemberCommandValidator(_regionsReadRepository.Object, _membersReadRepository.Object, _partnersReadRepository.Object);
 
             var result = await sut.TestValidateAsync(command);
